+++ ChangeLog.md
@@ -4,6 +4,7 @@
   rules. Output is sent to stdout, so you can save that output to file of
   your liking.
 - sudo: add `-A` option to use askpass and fail sooner if no sudo rules present
+- arcconf: check physical devices
 
 ## 3.0.3 (2013-11-12)
 - resync fixes